DOMA-2807 added settings PageContent component
