+++ apps/condo/domains/ticket/hooks/useTicketHintsTableColumns.tsx
@@ -1,15 +1,14 @@
-import get from 'lodash/get'
 import { useRouter } from 'next/router'
 import React, { useCallback, useMemo } from 'react'
-import { useIntl } from '../../../../../packages/@core.next/intl'
-import { htmlFor } from '../../common/components/containers/BehaviorRecorder'
-import { getTableCellRenderer } from '../../common/components/Table/Renders'
-import { getFilterIcon } from '../../common/components/TableFilter'
-import { FiltersMeta, getFilterDropdownByKey } from '../../common/utils/filters.utils'
-import { getFilteredValue } from '../../common/utils/helpers'
-import { parseQuery } from '../../common/utils/tables.utils'
-import { IFilters } from '../../contact/utils/helpers'
-import { getAddressRender } from '../../division/utils/clientSchema/Renders'
+import { useIntl } from '@core/next/intl'
+
+import { getTableCellRenderer } from '@condo/domains/common/components/Table/Renders'
+import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
+import { FiltersMeta, getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
+import { getFilteredValue } from '@condo/domains/common/utils/helpers'
+import { parseQuery } from '@condo/domains/common/utils/tables.utils'
+import { IFilters } from '@condo/domains/contact/utils/helpers'
+import { getAddressRender } from '@condo/domains/division/utils/clientSchema/Renders'
 
 export function useTicketHintsTableColumns <T> (filterMetas: Array<FiltersMeta<T>>) {
     const intl = useIntl()
